Pinecone/requirements.txt
streamlit 
PyPDF2 
langchain-openai 
langchain-pinecone 
pinecone-client 
openai
